myfirst/requirements.txt
aiogram==2.10.1
aiohttp==3.6.2
apiai==1.2.3
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
async-timeout==3.0.1
asyncio==3.4.3
attrs==20.2.0
autopep8==1.5.4
Babel==2.8.0
beautifulsoup4==4.9.1
bottle==0.12.18
bottle-websocket==0.2.9
bs4==0.0.1
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cryptography==3.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
Deprecated==1.2.10
dialogflow==1.0.0
discord==1.0.1
discord.py==1.4.1
distlib==0.3.1
distro==1.5.0
Django==3.1.1
django-grappelli==2.14.2
django-rest-framework==0.1.0
djangorestframework==3.12.1
docutils==0.16
Eel==0.14.0
filelock==3.0.12
flake8==3.8.3
Flask==1.1.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==0.16
flup==1.0.3
future==0.18.2
gevent==20.6.2
gevent-websocket==0.10.1
google-api-core==1.22.1
google-api-python-client==1.10.0
google-auth==1.20.1
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
greenlet==0.4.16
grpcio==1.31.0
html-parser==0.2
html5lib==1.1
httplib2==0.18.1
idna==2.10
image==1.5.32
isort==5.5.3
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
kivy-deps.glew==0.2.0
kivy-deps.sdl2==0.2.0
kiwisolver==1.2.0
lamson==1.3.4
langdetect==1.0.8
lazy-object-proxy==1.4.3
lxml==4.5.2
MarkupSafe==1.1.1
matplotlib==3.3.1
mccabe==0.6.1
multidict==4.7.6
numpy==1.19.1
oauth2client==4.1.3
parso==0.7.1
pbr==5.5.0
Pillow==7.2.0
pluggy==0.13.1
ply==3.11
protobuf==3.13.0
psutil==5.7.2
py-vkontakte==5.70.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pydocstyle==5.1.1
pyflakes==2.2.0
pygame==1.9.6
Pygments==2.7.1
pylint==2.6.0
PyNaCl==1.3.0
pyparsing==2.4.7
pypiwin32==223
PyQt5==5.15.0
PyQt5-sip==12.8.0
pyqt5-tools==5.15.0.1.7
PySide2==5.15.0
pyTelegramBotAPI==3.7.2
pytesseract==0.3.6
python-dateutil==2.8.1
python-dotenv==0.14.0
python-jsonrpc-server==0.4.0
python-language-server==0.35.1
python-telegram-bot==12.8
python3-openid==3.2.0
pytz==2020.1
pywin32==228
requests==2.24.0
rope==0.17.0
rsa==4.6
shiboken2==5.15.0
six==1.15.0
snowballstemmer==2.0.0
soupsieve==2.0.1
SQLAlchemy==1.3.19
sqlalchemy-migrate==0.13.0
sqlparse==0.3.1
telegram==0.0.1
Tempita==0.5.2
toml==0.10.1
tornado==6.0.4
ujson==3.2.0
uritemplate==3.0.1
urllib3==1.25.10
uuid==1.30
virtualenv==20.0.31
vk-api==11.8.0
webencodings==0.5.1
Werkzeug==1.0.1
whichcraft==0.6.1
wikipedia==1.4.0
Wikipedia-API==0.5.4
wrapt==1.12.1
yandex-checkout==1.6.4
yapf==0.30.0
yarl==1.5.1
zope.event==4.4
zope.interface==5.1.0